[HACK] Hack-in dynamic dimension for gradient

+++ frontend/catalyst/jax_primitives.py
@@ -354,7 +354,7 @@ def _grad_abstract(*args, jaxpr, fn, grad_params):
     # transformed_signature = calculate_grad_shape(signature, new_argnum)
     # return tuple(transformed_signature.get_results())
     t = [
-        (ShapedArray(shape=(), dtype=np.dtype(float)), False),
+        (ShapedArray(shape=(), dtype=np.dtype(int)), False),
         (DShapedArray(shape=(OutDBIdx(0),), dtype=np.dtype(float)), True)
     ]
     print("TTTTTTTT", t)
